[OPP-1378] legge til unleashproxyswitcher i forkant av dialog-implementasjonene
vi kan nå bytte mellom henvendelse og SF vha en enkel featuretoggel
