+++ web/src/main/java/no/nav/modiapersonoversikt/rest/api/Mapper.kt
@@ -5,8 +5,8 @@ import no.nav.modiapersonoversikt.legacy.api.domain.Saksbehandler
 import no.nav.modiapersonoversikt.legacy.api.domain.henvendelse.Melding
 import no.nav.modiapersonoversikt.legacy.sporsmalogsvar.consumer.henvendelse.domain.Traad
 import no.nav.modiapersonoversikt.rest.DATO_TID_FORMAT
-import no.nav.modiapersonoversikt.rest.dialog.MeldingDTO
-import no.nav.modiapersonoversikt.rest.dialog.TraadDTO
+import no.nav.modiapersonoversikt.rest.dialog.apis.MeldingDTO
+import no.nav.modiapersonoversikt.rest.dialog.apis.TraadDTO
 import org.joda.time.format.DateTimeFormat
 import kotlin.collections.HashMap
