+++ web/src/main/java/no/nav/modiapersonoversikt/rest/dialog/salesforce/SfLegacyDelsvarController.kt
@@ -1,0 +1,17 @@
+package no.nav.modiapersonoversikt.rest.dialog.salesforce
+
+import no.nav.modiapersonoversikt.rest.dialog.apis.DelsvarRestRequest
+import no.nav.modiapersonoversikt.rest.dialog.apis.DialogDelsvarApi
+import org.springframework.http.ResponseEntity
+import javax.servlet.http.HttpServletRequest
+import javax.ws.rs.NotSupportedException
+
+class SfLegacyDelsvarController : DialogDelsvarApi {
+    override fun svarDelvis(
+        httpRequest: HttpServletRequest,
+        fnr: String,
+        request: DelsvarRestRequest
+    ): ResponseEntity<Void> {
+        throw NotSupportedException("Delvar er ikke støttet av Salesforce")
+    }
+}
