+++ web/src/main/java/no/nav/modiapersonoversikt/service/unleash/Feature.kt
@@ -1,5 +1,6 @@
 package no.nav.modiapersonoversikt.service.unleash
 
 enum class Feature(val propertyKey: String) {
-    SAMPLE_FEATURE("feature.samplerfeature")
+    SAMPLE_FEATURE("feature.samplerfeature"),
+    USE_SALESFORCE_DIALOG("modiabrukerdialog.bruker-salesforce-dialoger")
 }
